+++ app/administration/ground-stations/actions.tsx
@@ -11,7 +11,7 @@ import {
 import { MoreHorizontal, Activity } from "lucide-react";
 import { useState } from "react";
 import { toast } from "sonner";
-import type { GroundStation } from "@/app/api/platform/ground-stations/mock";
+import type { GroundStation } from "@/app/api/platform/ground-stations/types";
 import { checkGroundStationHealth } from "@/app/api/platform/ground-stations/ground-station-service";
 import { EditGroundStationModal } from "./edit-ground-station-modal";
 import { DeleteGroundStationModal } from "./delete-ground-station-modal";
