+++ app/administration/ground-stations/delete-ground-station-modal.tsx
@@ -9,7 +9,7 @@ import {
   DialogHeader,
   DialogTitle,
 } from "@/components/ui/dialog";
-import type { GroundStation } from "@/app/api/platform/ground-stations/mock";
+import type { GroundStation } from "@/app/api/platform/ground-stations/types";
 import { deleteGroundStation } from "@/app/api/platform/ground-stations/ground-station-service";
 import { refreshGroundStations } from "@/app/actions/ground-stations";
 import { useState } from "react";
