--- app/api/authenticate.ts
@@ -1,9 +1,0 @@
-/* import { cookies } from "next/headers";
-import { NextResponse } from "next/server"; */
-
-// 🤬 I need to have shit compile
-export async function POST(/* request: Request */) {
-/*   const store = await cookies();
-  store.set("");
-  const res = NextResponse.co(); */
-}
