+++ app/globals.css
@@ -142,7 +142,7 @@ html, body {
     --sidebar-primary:  0 0% 9%;
     --sidebar-primary-foreground:  0 0% 98%;
     --sidebar-accent: 0 0% 96.1%;
-    --sidebar-accent-foreground: 0 0% 98%;
+    --sidebar-accent-foreground: 0 0% 9%;
     --sidebar-border: 0 0% 89.8%;
     --sidebar-ring: 0 0% 3.9%;
   }
