+++ app/platform/flight/flight-table.tsx
@@ -39,12 +39,9 @@ import {
   TooltipTrigger,
 } from "@/components/ui/tooltip";
 import { useRouter, useSearchParams } from "next/navigation";
-import {
-  FlightPlan,
-  FlightPlanStatus,
-} from "@/app/api/platform/flight/flight-plan-service";
-import { Satellite } from "@/app/api/platform/satellites/types";
-import { GroundStation } from "@/app/api/platform/ground-stations/mock";
+import type { FlightPlan, FlightPlanStatus } from "@/app/api/platform/flight/types";
+import type { Satellite } from "@/app/api/platform/satellites/types";
+import type { GroundStation } from "@/app/api/platform/ground-stations/types";
 import TablePagination from "@/components/ui/table-pagination";
 
 // Extended flight plan type with lookup data
