+++ app/platform/flight/search-form.tsx
@@ -27,12 +27,9 @@ import {
 } from "lucide-react";
 import { usePathname, useRouter, useSearchParams } from "next/navigation";
 import { useDebouncedCallback } from "use-debounce";
-import {
-  FlightPlan,
-  FlightPlanStatus,
-} from "@/app/api/platform/flight/flight-plan-service";
-import { Satellite } from "@/app/api/platform/satellites/types";
-import { GroundStation } from "@/app/api/platform/ground-stations/mock";
+import type { FlightPlan, FlightPlanStatus } from "@/app/api/platform/flight/types";
+import type { Satellite } from "@/app/api/platform/satellites/types";
+import type { GroundStation } from "@/app/api/platform/ground-stations/types";
 import { cn } from "@/lib/utils";
 
 interface SearchFormProps {
