--- lib/state.ts
@@ -1,1 +1,0 @@
-"use client";
